refactor: separate classes and interfaces into new files
